backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
cohere==5.5.8
python-dotenv==1.0.1
pydantic==2.8.2
orjson==3.10.7
